# clients/shared/openapi.yaml
openapi: 3.0.3
info:
  title: Called It API
  version: 1.0.0
  description: >
    Backend contract for **Called It** — a daily casual-forecasting game. Every day a single
    synchronized global set of binary questions "drops"; players lock in a pick (Side A / Side B)
    or Skip before a hard 6-hour lock, questions auto-resolve, and independent per-category streaks
    plus a never-resetting Total Score drive the leaderboards.

    The invite-only TEST POC adds explicitly simulated, shared two-minute rounds behind
    TestMode:Enabled. Production phone/social authentication and six-hour daily sets are unchanged.
    The shared Expo/React Native client uses the TEST endpoints for phone-browser play.
    This document also remains the contract for the existing iOS client.
servers:
  - url: https://{apiFqdn}
    description: Deployed API (Azure Container Apps)
    variables:
      apiFqdn:
        default: called-it-dev.<region>.azurecontainerapps.io
  - url: http://localhost:5080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/test/login:
    post:
      tags: [Test]
      summary: Sign in with a private TEST invite; never a phone or social account
      description: >
        Only available with TestMode:Enabled in Test or Development. Each configured stable invite
        ID owns one durable, non-admin, non-contact-discoverable account. Display names are not
        identifiers and may be shared. The raw invite is never persisted or logged.
        Limited to 20 attempts per remote IP per five-minute server-process window, no queue.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TestLoginRequest' }
      responses:
        '200':
          description: Authenticated TEST account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '403':
          description: Invalid invite code
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '404': { description: TEST mode disabled }
        '413': { description: Request body exceeds 4096 bytes }
        '429': { $ref: '#/components/responses/Problem' }
  /api/test/game:
    get:
      tags: [Test]
      summary: Advance shared TEST rounds and read the game, previous results, and personal stats
      description: >
        Requires an authenticated, still-configured TEST invite. Resolves the expired active
        round transactionally, recomputes real scores, and creates one new 120-second round at
        the request time. Idle periods are not backfilled. Each round has one clearly labeled
        sample Sports, Finance, and Pop Culture question. Results are randomly simulated only
        after lock and never disclosed ahead of time. Times and submission locks are authoritative
        on the server. Only the caller's picks/skips are returned.
      responses:
        '200':
          description: Current TEST game
          headers:
            Cache-Control:
              schema: { type: string }
              description: Clients must not cache authenticated game snapshots.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestGameView' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { description: TEST mode disabled }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Problem:
      description: Error (RFC 7807 problem details)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
    Side:
      type: string
      enum: [A, B]
      description: The two sides of a binary question.
    Outcome:
      type: string
      enum: [Unresolved, SideA, SideB, Void]
...
    TestLoginRequest:
      type: object
      required: [inviteCode, displayName]
      properties:
        inviteCode:
          type: string
          minLength: 32
          maxLength: 256
          pattern: '^\S+$'
          writeOnly: true
          description: >
            Strong random invite (at least 128 bits of randomness; recommend 32 random bytes
            encoded as base64url). Compared exactly as UTF-8; no whitespace or trimming.
        displayName:
          type: string
          minLength: 1
          maxLength: 60
          description: Trimmed before validation; must not be blank or contain control characters.
...
    AuthResult:
      type: object
      required: [accessToken, accessTokenExpiresAt, refreshToken, userId, displayName, isAdmin]
      properties:
        accessToken: { type: string }
        accessTokenExpiresAt: { type: string, format: date-time }
        refreshToken: { type: string }
        userId: { type: string, format: uuid }
        displayName: { type: string }
        isAdmin: { type: boolean }
...
    DailySetQuestionView:
      type: object
      required: [questionId, categoryCode, categoryName, text, sideALabel, sideBLabel, myPick, mySkip, outcome]
      properties:
        questionId: { type: string, format: uuid }
        categoryCode: { type: string, example: pop_culture }
        categoryName: { type: string, example: Pop Culture }
        text: { type: string }
        sideALabel: { type: string }
        sideBLabel: { type: string }
        myPick:
          allOf: [{ $ref: '#/components/schemas/Side' }]
          nullable: true
        mySkip: { type: boolean }
        outcome: { $ref: '#/components/schemas/Outcome' }
    DailySetView:
      type: object
      required: [id, dropAtUtc, locksAtUtc, isOpen, questions]
      properties:
        id: { type: string, format: uuid }
        dropAtUtc: { type: string, format: date-time }
        locksAtUtc: { type: string, format: date-time }
        isOpen: { type: boolean }
        questions:
          type: array
          items: { $ref: '#/components/schemas/DailySetQuestionView' }
    TestGameView:
      type: object
      required: [serverTimeUtc, currentRound, previousRound, stats]
      properties:
        serverTimeUtc: { type: string, format: date-time }
        currentRound:
          allOf: [{ $ref: '#/components/schemas/DailySetView' }]
          description: Always present on success. Current outcomes remain Unresolved before lock.
        previousRound:
          allOf: [{ $ref: '#/components/schemas/DailySetView' }]
          nullable: true
          description: Most recently completed simulated round, or null before the first completion.
        stats: { $ref: '#/components/schemas/TestPlayerStats' }
    TestPlayerStats:
      type: object
      required: [totalScore, overallStreak, categories]
      properties:
        totalScore: { type: integer, format: int32, minimum: 0 }
        overallStreak:
          type: integer
          format: int32
          minimum: 0
          description: Sum of current per-category streaks.
        categories:
          type: array
          items: { $ref: '#/components/schemas/TestCategoryStats' }
    TestCategoryStats:
      type: object
      required: [categoryCode, categoryName, currentStreak, bestStreak, totalCorrect]
      properties:
        categoryCode: { type: string, enum: [sports, finance, pop_culture] }
        categoryName: { type: string }
        currentStreak: { type: integer, format: int32, minimum: 0 }
        bestStreak: { type: integer, format: int32, minimum: 0 }
        totalCorrect: { type: integer, format: int32, minimum: 0 }
...
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer, format: int32 }
        detail: { type: string }
        instance: { type: string }